scst_mem: Remove a superfluous cast
git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@6843 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/src/scst_mem.c
@@ -366,7 +366,7 @@ static unsigned long __sgv_can_be_shrunk(void)
 	}
 	spin_unlock_bh(&sgv_pools_lock);
 
-	res = max((int)0, inactive_pages - sgv_lo_wmk);
+	res = max(0, inactive_pages - sgv_lo_wmk);
 	TRACE_MEM("Can free %ld (total %d)", res, atomic_read(&sgv_pages_total));
 
 	TRACE_EXIT_RES(res);
